+++ HISTORY.rst
@@ -9,6 +9,8 @@
 
 - 357_: added psutil.Process.cpu_num() (what CPU a process is on).
 - 941_: added psutil.cpu_freq() (CPU frequency).
+- 956_: cpu_affinity([]) can now be used as an alias to set affinity against
+  all eligible CPUs.
 
 **Bug fixes**